+++ .storybook/manager.js
@@ -1,0 +1,6 @@
+import { addons } from '@storybook/addons';
+import fivebim from './theme';
+
+addons.setConfig({
+    theme: fivebim,
+});