+++ .storybook/preview.js
@@ -1,7 +1,17 @@
 import "@fortawesome/react-fontawesome";
 import "@fortawesome/free-solid-svg-icons";
+
 export const parameters = {
   actions: { argTypesRegex: "^on[A-Z].*" },
+  backgrounds: {
+    default: 'default',
+    values: [
+      {
+        name: 'default',
+        value: '#ffffff',
+      },
+    ],
+  },
   controls: {
     matchers: {
       color: /(background|color)$/i,