+++ .storybook/theme.js
@@ -1,0 +1,8 @@
+import { create } from '@storybook/theming';
+
+export default create({
+  base: 'light',
+  brandTitle: 'Tabla FiveBim',
+  brandImage: './images/logo.png',
+  brandTarget: '_self',
+});